% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval-metrics.R
\name{eval_child_parent_dir}
\alias{eval_child_parent_dir}
\title{Evaluation metric (child -> parent)}
\usage{
eval_child_parent_dir(r_child, r_parent)
}
\arguments{
\item{r_child}{Raster processed with a "smaller" weight matrix.}

\item{r_parent}{Raster processed with a "larger" weight matrix.}
}
\value{
A number between 0 and 1
}
\description{
Evaluation metric (child -> parent)
}
\author{
Viliam Simko
}
