% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetisOrd.R
\name{GetisOrd}
\alias{GetisOrd}
\title{Getis-Ord G* statistics using a weight matrix.
Data on the border are padded with the mean value during convolution.}
\usage{
GetisOrd(r, w, st = GetisOrdStandardStats(r))
}
\arguments{
\item{r}{Input raster}

\item{w}{Square weight matrix with odd number of rows/columns.}

\item{st}{Expecting a list of local statistics, see
\link{GetisOrdStandardStats} and \link{GetisOrdLocalStats}}
}
\value{
New raster of the same size as the original input raster \code{r}.
}
\description{
Getis-Ord G* statistics using a weight matrix.
Data on the border are padded with the mean value during convolution.
}

