% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetisOrdStandardStats.R
\name{GetisOrdStandardStats}
\alias{GetisOrdStandardStats}
\title{Standard Getis-Ord G* statistics.}
\usage{
GetisOrdStandardStats(r)
}
\arguments{
\item{r}{Input raster layer}
}
\value{
Returns a list containing:
\item{rmean}{Global mean of all raster cells.}
\item{n}{Number of all cells within the raster.}
\item{SD}{Standard deviation of all raster cells (single value)}
\item{MEAN}{Same as rmean in standard Getis-Ord}
}
\description{
Standard Getis-Ord G* statistics.
}

