% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colortable-addons.R
\name{interpolate_colortable}
\alias{interpolate_colortable}
\title{Helper functon to generates a color table.}
\usage{
interpolate_colortable(howmany, ...)
}
\arguments{
\item{howmany}{Number of elements of the resulting vector.}

\item{...}{A vecor of colors to be used for interpolation.}
}
\value{
Returns a vector representing the color table.
}
\description{
Helper functon to generates a color table.
}
\examples{
ct <- interpolate_colortable(10, "red", "green" ,"blue")
plot_color_table(ct)
}

