% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspot_analysis_lib.R
\name{JustConvolve}
\alias{JustConvolve}
\title{Focal convolution of a raster using a weight matrix.
Data on the border are padded with the mean value during convolution.}
\usage{
JustConvolve(r, w)
}
\arguments{
\item{r}{raster}

\item{w}{square weight matrix with odd number of rows/columns}
}
\value{
convoluted raster
}
\description{
Focal convolution of a raster using a weight matrix.
Data on the border are padded with the mean value during convolution.
}
\examples{
r <- raster(matrix(rnorm(10000), 100, 100))
w <- weight_matrix_circular(9)
plot(JustConvolve(r, w))
}

