% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspot_analysis_lib.R
\name{plot_color_table}
\alias{plot_color_table}
\title{Plots a color table (useful for debugging).
Uses \link{raster} for plotting.}
\usage{
plot_color_table(ct)
}
\arguments{
\item{ct}{Vector of colors}
}
\description{
Plots a color table (useful for debugging).
Uses \link{raster} for plotting.
}
\examples{
plot_color_table(c("black", "red", "gold")) # German flag
}

