% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_temp_map.R
\name{plot_temp_map}
\alias{plot_temp_map}
\title{This function is suitable for plotting raw temperatures encoded as a raster.}
\usage{
plot_temp_map(r, title, colorscale_title = "", contour_thresh = NULL,
  contour_quantiles = NULL, contour_col = c("blue", "red"),
  ctable = interpolate_colortable(howmany = 100, "blue", "green", "yellow",
  "red"), topn = 0, bottomn = 0)
}
\arguments{
\item{r}{Input raster containing temperature values}

\item{title}{Title for the plot}

\item{colorscale_title}{Title for the color legend}
}
\description{
This function is suitable for plotting raw temperatures encoded as a raster.
}
\examples{
r <- raster(matrix(rnorm(400), 20))
plot_temp_map(r, "Random temperatures")

}

