% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight-matrix.R
\name{plot_weight_matrix}
\alias{plot_weight_matrix}
\title{Renders a plot showing the given weight matrix.}
\usage{
plot_weight_matrix(w, show.legend = TRUE)
}
\arguments{
\item{w}{Weight matrix}

\item{show.legend}{When TRUE, a legend is also rendered within the plot.}
}
\description{
Renders a plot showing the given weight matrix.
}
\examples{
w <- weight_matrix_circular_fade(11,2)
plot_weight_matrix(w)

}
\author{
Viliam Simko
}
