% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_zscore.R
\name{plot_zscore}
\alias{plot_zscore}
\title{This function is suitable for plotting zscores encoded as a raster.}
\usage{
plot_zscore(r, sigma_thresh = 5, hotspot_quantile_thresh = 0.01,
  title = NULL)
}
\arguments{
\item{r}{Raster to be plotted.}

\item{sigma_thresh}{Sigma (zscore) value used for filtering significant
raster cells.}

\item{hotspot_quantile_thresh}{Quantile value used as a threshold for
plotting hotspots.}

\item{title}{Title of the plot}
}
\description{
This function is suitable for plotting zscores encoded as a raster.
}
\author{
Viliam Simko
}
