% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotspot_analysis_lib.R
\name{select_extent_manually}
\alias{select_extent_manually}
\title{Helper function that can be used for picking the extend manually from
a plotted raster. (use mouse to pick coordinates on the map)}
\usage{
select_extent_manually(r)
}
\arguments{
\item{r}{raster}
}
\value{
\link{Extent} object representing the selected area
}
\description{
Helper function that can be used for picking the extend manually from
a plotted raster. (use mouse to pick coordinates on the map)
}

