% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightMatrix.R
\name{sigmoid}
\alias{sigmoid}
\title{Helper function used in \code{sigmoid_falloff}. See example.}
\usage{
sigmoid(x)
}
\arguments{
\item{x}{X-coordinate.}
}
\value{
value on the sigmoid curve
}
\description{
Helper function used in \code{sigmoid_falloff}. See example.
}
\examples{
plot(sigmoid(-10:10), type="l")
}

