% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightMatrix.R
\name{sigmoid_falloff}
\alias{sigmoid_falloff}
\title{Helper function which produces a smooth transition from 1 to 0 with a
sigmoidal falloff. See example.}
\usage{
sigmoid_falloff(x, size, fsize)
}
\arguments{
\item{x}{Input number on x-axis}

\item{size}{Size of the interpolation}

\item{fsize}{size of the fallof part}
}
\description{
Helper function which produces a smooth transition from 1 to 0 with a
sigmoidal falloff. See example.
}
\examples{
x <- 1:100
y <- sigmoid_falloff(x, 100, 20)
plot(x, y)
}

