% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight-matrix.R
\name{weight_matrix_circular}
\alias{weight_matrix_circular}
\title{Circular shape with a sharp falloff at the edge.}
\usage{
weight_matrix_circular(size)
}
\arguments{
\item{size}{Odd number representing the matrix width and height.
Minimal matrix size is 3x3.}
}
\value{
boolean square matrix
}
\description{
Circular shape with a sharp falloff at the edge.
}
\examples{
w <- weight_matrix_circular(5)
plot_weight_matrix(w)
}
\author{
Viliam Simko
}
