% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightMatrix.R
\name{weight_matrix_circular_fade}
\alias{weight_matrix_circular_fade}
\title{Circular shape with sigmoid falloff at the edge.}
\usage{
weight_matrix_circular_fade(size, fsize)
}
\arguments{
\item{size}{Odd number representing the matrix width and height.
Minimal matrix size is 3x3.}
}
\value{
square matrix of doubles
}
\description{
Circular shape with sigmoid falloff at the edge.
}
\examples{
w <- weight_matrix_circular_fade(21, 3)
plot_weight_matrix(w)
}

