% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WeightMatrix.R
\name{weight_matrix_squared}
\alias{weight_matrix_squared}
\title{Just a simple square matrix.}
\usage{
weight_matrix_squared(size)
}
\arguments{
\item{size}{Odd number representing the matrix width and height. Minimal
matrix size is 3x3.}
}
\value{
boolean square matrix
}
\description{
Just a simple square matrix.
}
\examples{
weight_matrix_squared(5)
}

