% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval-metrics.R
\name{eval_child_parent_dir}
\alias{eval_child_parent_dir}
\title{Evaluation metric (child -> parent)}
\usage{
eval_child_parent_dir(r_child, r_parent)
}
\description{
Evaluation metric (child -> parent)
}

