% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetisOrdLocalStats.R
\name{GetisOrdLocalStats}
\alias{GetisOrdLocalStats}
\title{Locality-aware Getis-Ord G* statistics.}
\usage{
GetisOrdLocalStats(r, focus_matrix)
}
\arguments{
\item{r}{input raster}

\item{focus}{Square matrix with odd number of rows/column and boolean cells.}
}
\value{
Returns a list containing:
\item{rmean}{Global mean of all raster cells.}
\item{n}{Number of all cells within the raster.}
\item{SD}{A raster where each cell corresponds to a focal standard deviation
from area around the cell given by the matrix \code{focus}}
\item{MEAN}{A raster where each cell corresponds to a focal mean from area
around the cell given by the matrix \code{focus}}
\item{focus}{Copy of the \code{focus} matrix given as input parameter.}
}
\description{
Locality-aware Getis-Ord G* statistics.
}

